+++ src/Javinizer/Private/Scraper.R18.ps1
@@ -246,18 +246,6 @@ function Get-R18Series {
     }
 }
 
-function Get-R18Rating {
-    param (
-        [Parameter(Mandatory = $true, Position = 0, ValueFromPipeline = $true)]
-        [Object]$Webrequest
-    )
-
-    process {
-        $rating = ''
-        Write-Output $rating
-    }
-}
-
 function Get-R18Genre {
     param (
         [Parameter(Mandatory = $true, Position = 0, ValueFromPipeline = $true)]
@@ -331,14 +319,14 @@ function Get-R18Actress {
                 }
             } else {
                 try {
-                    $japaneseActressName = ((Invoke-WebRequest -Uri ($zhActressUrl -replace 'lg=en', 'lg=zh')).Content | Select-String -Pattern '<h1 class="txt01">(.*)<\/h1><\/div>').Matches.Groups[1].Value
+                    $jaActressName = ((Invoke-WebRequest -Uri ($zhActressUrl -replace 'lg=en', 'lg=zh')).Content | Select-String -Pattern '<h1 class="txt01">(.*)<\/h1><\/div>').Matches.Groups[1].Value
                 } catch {
                     return
                 }
                 $movieActressObject += [PSCustomObject]@{
                     LastName     = ($actressName -split ' ')[1]
                     FirstName    = ($actressName -split ' ')[0]
-                    JapaneseName = $japaneseActressName
+                    JapaneseName = $jaActressName
                     ThumbUrl     = $thumbUrl
                 }
             }