+++ src/Javinizer/Public/Get-Jav321Data.ps1
@@ -18,19 +18,20 @@ function Get-Jav321Data {
         }
 
         $movieDataObject = [PSCustomObject]@{
-            Source          = 'jav321'
-            Url             = $Url
-            Id              = Get-Jav321Id -WebRequest $webRequest
-            Title           = Get-Jav321Title -WebRequest $webRequest
-            ReleaseDate     = Get-Jav321ReleaseDate -WebRequest $webRequest
-            ReleaseYear     = Get-Jav321ReleaseYear -WebRequest $webRequest
-            Runtime         = Get-Jav321Runtime -WebRequest $webRequest
-            Maker           = Get-Jav321Maker -WebRequest $webRequest
-            Actress         = (Get-Jav321Actress -WebRequest $webRequest).Name
-            ActressThumbUrl = (Get-Jav321Actress -WebRequest $webRequest).ThumbUrl
-            Genre           = Get-Jav321Genre -WebRequest $webRequest
-            CoverUrl        = Get-Jav321CoverUrl -WebRequest $webRequest
-            ScreenshotUrl   = Get-Jav321ScreenshotUrl -WebRequest $webRequest
+            Source        = 'jav321ja'
+            Url           = $Url
+            Id            = Get-Jav321Id -WebRequest $webRequest
+            Title         = Get-Jav321Title -WebRequest $webRequest
+            Description   = Get-Jav321Description -WebRequest $webRequest
+            ReleaseDate   = Get-Jav321ReleaseDate -WebRequest $webRequest
+            ReleaseYear   = Get-Jav321ReleaseYear -WebRequest $webRequest
+            Runtime       = Get-Jav321Runtime -WebRequest $webRequest
+            Series        = Get-Jav321Series -WebRequest $webRequest
+            Maker         = Get-Jav321Maker -WebRequest $webRequest
+            Actress       = Get-Jav321Actress -WebRequest $webRequest
+            Genre         = Get-Jav321Genre -WebRequest $webRequest
+            CoverUrl      = Get-Jav321CoverUrl -WebRequest $webRequest
+            ScreenshotUrl = Get-Jav321ScreenshotUrl -WebRequest $webRequest
         }
 
         Write-JVLog -Write:$script:JVLogWrite -LogPath $script:JVLogPath -WriteLevel $script:JVLogWriteLevel -Level Debug -Message "Jav321 data object: $($movieDataObject | ConvertTo-Json -Depth 32 -Compress)"