+++ src/Javinizer/Public/Javinizer.ps1
@@ -83,8 +83,8 @@ function Javinizer {
     .PARAMETER JavbusZh
         Specifies to search Javbus-Chinese when using -Find.
 
-    .PARAMETER Jav321
-        Specifies to search Jav321 when using -Find.
+    .PARAMETER Jav321Ja
+        Specifies to search Jav321-Japanese when using -Find.
 
     .PARAMETER SetEmbyThumbs
         Specifies to set Emby/Jellyfin actress thumbnails using the thumbnail csv. If 'location.thumbcsv' is not specified in the settings file,
@@ -212,6 +212,13 @@ function Javinizer {
     Updates existing sorted nfo files from a path with updated aliases, thumburls, names, ignored genres, and genre replacements according to the settings.
 
     .EXAMPLE
+    Javinizer -UpdateThumbs -Pages 1,10
+
+    Description
+    -----------
+    Updates the actress csv file using a range of pages to scrape from.
+
+    .EXAMPLE
     Javinizer -OpenSettings
 
     Description
@@ -242,6 +249,7 @@ function Javinizer {
         [Array]$Url,
 
         [Parameter(ParameterSetName = 'Path')]
+        [Parameter(ParameterSetName = 'Info')]
         [System.IO.FileInfo]$SettingsPath,
 
         [Parameter(ParameterSetName = 'Path')]
@@ -300,7 +308,7 @@ function Javinizer {
         [Switch]$JavbusZh,
 
         [Parameter(ParameterSetName = 'Info')]
-        [Switch]$Jav321,
+        [Switch]$Jav321Ja,
 
         [Parameter(ParameterSetName = 'Emby')]
         [Switch]$SetEmbyThumbs,
@@ -416,8 +424,6 @@ function Javinizer {
         if ($Set) {
             try {
                 foreach ($item in $Set.GetEnumerator()) {
-                    $settingName = $item.Key
-                    $settingValue = $item.Value
                     $Settings."$($item.Key)" = $item.Value
                     Write-JVLog -Write:$script:JVLogWrite -LogPath $script:JVLogPath -WriteLevel $script:JVLogWriteLevel -Level Debug -Message "[$($MyInvocation.MyCommand.Name)] [Setting - $($item.Key)] replaced as [$($item.Value)]"
                 }
@@ -460,7 +466,7 @@ function Javinizer {
                     }
                 } else {
                     $data = Get-JVData -Id $Find -R18:$R18 -R18Zh:$R18Zh -Javlibrary:$Javlibrary -JavlibraryJa:$JavlibraryJa -JavlibraryZh:$JavlibraryZh -Dmm:$Dmm `
-                        -Javbus:$Javbus -JavbusJa:$JavbusJa -JavbusZh:$JavbusZh -Jav321:$Jav321
+                        -Javbus:$Javbus -JavbusJa:$JavbusJa -JavbusZh:$JavbusZh -Jav321Ja:$Jav321Ja
                 }
 
                 if ($Aggregated) {